Presence pulse/requirements.txt
django>=4.2
mysql-connector-python>=8.0
gunicorn
fcntl